requirements.txt
aiohttp
jsonschema
python-dateutil
uvloop; sys_platform != 'win32'
